requirements.txt
# Core monitoring
psutil>=5.9.0
watchdog>=3.0.0

# Configuration
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0

# UI - Systray
pystray>=0.19.0
Pillow>=10.0.0
plyer>=2.1.0

# Web dashboard (Phase 2)
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0

# Machine Learning
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.3.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Windows-specific (optional)
pywin32>=306; sys_platform == 'win32'
